+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2026-05-06
+
+### Added
+- Nexus update key
+- New config RequireSonarBobber, makes the fish icon silhouette without sonar bobber.
+- New config BypassCatchLimit, having it off makes target bait respect catch limit
+
 ## [0.4.2] - 2026-05-04
 
 ### Fixed